requirements.txt
watchdog
pytesseract
PyQt5
pdf2image
Pillow
opencv-python
numpy
